#77 clean code

+++ src/main/java/dev/tobee/telegram/request/body/ForwardMessageBody.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record ForwardMessageBody(
         @JsonProperty("chat_id") long chatId,
-        @JsonProperty("from_chat_id") int fromChatId,
+        @JsonProperty("from_chat_id") long fromChatId,
         @JsonProperty("disable_notification") Optional<Boolean> disableNotification,
         @JsonProperty("message_id") OptionalLong messageId
 ) {
